backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
curl_cffi==0.15.0
fastapi==0.135.3
flatbuffers==25.12.19
frozendict==2.4.7
gast==0.7.0
google-pasta==0.2.0
grpcio==1.80.0
h11==0.16.0
h5py==3.14.0
idna==3.11
joblib==1.5.3
keras==3.14.0
libclang==18.1.1
llvmlite==0.44.0
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.4
multitasking==0.0.12
namex==0.1.0
numba==0.61.2
numpy==2.2.6
opt_einsum==3.4.0
optree==0.19.0
packaging==26.1
pandas==3.0.2
pandas-ta==0.4.71b0
peewee==4.0.4
platformdirs==4.9.6
protobuf==7.34.1
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.33.1
rich==15.0.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
starlette==1.0.0
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
websockets==16.0
wheel==0.46.3
wrapt==2.1.2
yfinance==1.2.2
PyJWT==2.12.1
